add icon-titles rule and specs
this commit adds a rule that checks to make sure any instance of jha-icon-* includes a title attribute.  Chad K mentioned that it would be a good idea to have it.

The implementation is pretty simple.. it just looks for a tag named `jha-icon-*` (uses regex to check the wildcard) and makes sure the title attribute is present.